requirements.txt
flask
ultralytics
pillow
gunicorn